+++ NEWS.md
@@ -1,5 +1,10 @@
 # dplyr (development version)
 
+* The `.keep`, `.before`, and `.after` arguments to `mutate()` 
+  are no longer experimental.
+  
+* The `rows_*()` family of functions are no longer experimental.
+
 * `desc()` gives a useful error message if you give it a non-vector (#6028).
 
 * `slice_sample()` returns a data frame or group with the same number of rows as