+++ NEWS.md
@@ -1,7 +1,8 @@
 # dplyr (development version)
 
 * `all_equal()` is formally deprecated. We've advised against it for
-  some time, and we explicitly recommend you use something else (#6324).
+  some time, and we explicitly recommend you use `all.equal()`,
+  manually reordering the rows and columns (#6324).
 
 * `slice()` helpers again produce output equivalent to `slice(.data, 0)` when
   the `n` or `prop` argument is 0, fixing a bug introduced in the previous